requirements.txt
awscrt == 0.12.1
awsiotsdk == 1.7.1
pip == 21.3.1
setuptools==58.3.0
wheel==0.37.0
